requirements.txt
streamlit
google-generativeai
whisper
gtts
pydub
sounddevice
torch
